+++ src/main/java/hu/ibello/plugins/PluginInitializer.java
@@ -24,6 +24,7 @@ import hu.ibello.data.TestDataBuilder;
 import hu.ibello.functions.RegressionTool;
 import hu.ibello.graph.GraphTool;
 import hu.ibello.output.TestResultLoader;
+import hu.ibello.requirements.ModuleHandler;
 import hu.ibello.requirements.RequirementHandler;
 import hu.ibello.table.TableTool;
 import hu.ibello.transform.CsvTransformer;
@@ -146,6 +147,12 @@ public interface PluginInitializer {
 	 * @return the requirement handler object
 	 */
 	public RequirementHandler requirements();
+	
+	/**
+	 * Returns an object which is designed to load module files from an ibello project.
+	 * @return the module handler object
+	 */
+	public ModuleHandler modules();
 	
 	/**
 	 * Returns an object which can load ibello tests results.