+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - The test result XML contains the category of the test steps
 - The `Step` class may contain `DATATABLE` type parameter
 - New `GraphTool.createLabeledGraph(...)` method to create labeled plot
+- New `PluginInitializer.modules()` method to load module information
 
 ## 1.21.2